# Conflicts:
#	example/controller-registration.yaml
#	go.mod
#	go.sum
#	pkg/azure/client/availabilityset.go
#	pkg/azure/client/factory.go
#	pkg/azure/client/helper.go
#	pkg/azure/client/mock/mocks.go
#	pkg/azure/client/natgateway.go
#	pkg/azure/client/networksecuritygroup.go
#	pkg/azure/client/resourcegroup.go
#	pkg/azure/client/routetables.go
#	pkg/azure/client/subnet.go
#	pkg/azure/client/types.go
#	pkg/azure/client/useridentity.go
#	pkg/azure/client/vnet.go
#	pkg/controller/backupbucket/actuator.go
#	pkg/controller/bastion/actuator.go
#	pkg/controller/bastion/actuator_reconcile.go
#	pkg/controller/bastion/bastion_test.go
#	pkg/controller/bastion/resourcesdefine.go
#	pkg/controller/dnsrecord/actuator_test.go
#	pkg/controller/infrastructure/actuator.go
#	pkg/controller/infrastructure/actuator_delete.go
#	pkg/controller/infrastructure/actuator_migrate.go
#	pkg/controller/infrastructure/actuator_reconcile.go
#	pkg/controller/infrastructure/actuator_test.go
#	pkg/controller/infrastructure/flowreconciler.go
#	pkg/controller/infrastructure/infraflow/azreconciler.go
#	pkg/controller/infrastructure/infraflow/azreconciler_test.go
#	pkg/controller/infrastructure/infraflow/mock_factory_test.go
#	pkg/controller/infrastructure/infraflow/reconciler.go
#	pkg/controller/infrastructure/infraflow/shared/basic_context_test.go
#	pkg/controller/infrastructure/infraflow/shared/tf_state_test.go
#	pkg/controller/infrastructure/infraflow/shared/whiteboard_test.go
#	pkg/controller/infrastructure/infraflow/terraformadapter.go
#	pkg/controller/infrastructure/infraflow/terraformadapter_test.go
#	pkg/controller/worker/machine_dependencies_test.go
#	pkg/controller/worker/machineset_vmo.go
#	pkg/internal/auth.go
#	pkg/internal/infrastructure/helper.go
#	pkg/internal/infrastructure/helper_test.go
#	pkg/internal/infrastructure/terraform.go
#	test/integration/bastion/bastion_test.go
#	vendor/modules.txt

+++ pkg/controller/backupbucket/actuator.go
@@ -17,12 +17,14 @@ package backupbucket
 import (
 	"context"
 
-	azureclient "github.com/gardener/gardener-extension-provider-azure/pkg/azure/client"
-	"github.com/go-logr/logr"
-
 	"github.com/gardener/gardener/extensions/pkg/controller/backupbucket"
+	"github.com/gardener/gardener/extensions/pkg/util"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/go-logr/logr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/gardener/gardener-extension-provider-azure/pkg/apis/azure/helper"
+	azureclient "github.com/gardener/gardener-extension-provider-azure/pkg/azure/client"
 )
 
 type actuator struct {
@@ -50,19 +52,19 @@ func (a *actuator) Reconcile(ctx context.Context, _ logr.Logger, backupBucket *e
 	if backupBucket.Status.GeneratedSecretRef == nil {
 		storageAccountName, storageAccountKey, err := ensureBackupBucket(ctx, a.client, factory, backupBucket)
 		if err != nil {
-			return err
+			return util.DetermineError(err, helper.KnownCodes)
 		}
 		// Create the generated backupbucket secret.
 		if err := a.createBackupBucketGeneratedSecret(ctx, backupBucket, storageAccountName, storageAccountKey); err != nil {
-			return err
+			return util.DetermineError(err, helper.KnownCodes)
 		}
 	}
 
 	storageClient, err := factory.Storage(ctx, *backupBucket.Status.GeneratedSecretRef)
 	if err != nil {
-		return err
+		return util.DetermineError(err, helper.KnownCodes)
 	}
-	return storageClient.CreateContainerIfNotExists(ctx, backupBucket.Name)
+	return util.DetermineError(storageClient.CreateContainerIfNotExists(ctx, backupBucket.Name), helper.KnownCodes)
 }
 
 func (a *actuator) Delete(ctx context.Context, _ logr.Logger, backupBucket *extensionsv1alpha1.BackupBucket) error {
@@ -77,22 +79,28 @@ func (a *actuator) Delete(ctx context.Context, _ logr.Logger, backupBucket *exte
 		return err
 	}
 
-	// Get a storage account client to delete the backup container in the storage account.
-	storageClient, err := factory.Storage(ctx, *backupBucket.Status.GeneratedSecretRef)
+	secret, err := a.getBackupBucketGeneratedSecret(ctx, backupBucket)
 	if err != nil {
-		return err
+		return util.DetermineError(err, helper.KnownCodes)
 	}
-	if err := storageClient.DeleteContainerIfExists(ctx, backupBucket.Name); err != nil {
-		return err
+	if secret != nil {
+		// Get a storage account client to delete the backup container in the storage account.
+		storageClient, err := factory.Storage(ctx, *backupBucket.Status.GeneratedSecretRef)
+		if err != nil {
+			return util.DetermineError(err, helper.KnownCodes)
+		}
+		if err := storageClient.DeleteContainerIfExists(ctx, backupBucket.Name); err != nil {
+			return util.DetermineError(err, helper.KnownCodes)
+		}
 	}
 
 	// Get resource group client and delete the resource group which contains the backup storage account.
 	groupClient, err := factory.Group()
 	if err != nil {
-		return err
+		return util.DetermineError(err, helper.KnownCodes)
 	}
-	if err := groupClient.DeleteIfExists(ctx, backupBucket.Name); err != nil {
-		return err
+	if err := groupClient.Delete(ctx, backupBucket.Name); err != nil {
+		return util.DetermineError(err, helper.KnownCodes)
 	}
 
 	// Delete the generated backup secret in the garden namespace.
